requirements.txt
torch
torchvision
transformers
efficientnet-pytorch
opencv-python
numpy
pillow
fastapi
uvicorn
python-multipart
reportlab